fix: Move classes to shared folder
